Refactor OG generation as a plugin
