+++ src/_includes/layouts/base.njk
@@ -18,13 +18,12 @@
       property="og:url"
       content="{{ page.url | htmlBaseUrl(metadata.url) }}"
     />
-    {% set ogImage = '/images/lyza.gif' | htmlBaseUrl(metadata.url) %}
     {% if og and og.image %}
       {% set ogImage = og.image %}
     {% endif %}
     <meta
       property="og:image"
-      content="{{ title | cloudinaryOGImage(ogImage) | default(defaultOgImage) }}"
+      content="{% cloudinaryOGImage title, ogImage %}"
     />
     {% if excerpt %}
       <meta property="og:description" content="{{ excerpt }}" />
